refactor: move upgrade helper functions back to helpers.py
The prior split (upgrade_tests_helper.py) moved 6 general-purpose
helper functions out of helpers.py alongside the 5 run_*_upgrade_scenario
orchestrators, which meant helpers.py itself showed a diff against
main for functions that were only relocated, not changed.

Restored to helpers.py (now byte-identical to main again):
assert_traefik_revision, verify_http_on_all_units, verify_http_on_unit,
leader_unit_name, force_leader_change, bring_up_traefik_without_certificate_provider.

Renamed upgrade_tests_helper.py -> upgrade_tests_scenarios.py, keeping
only the 5 run_*_upgrade_scenario orchestrators, which now import the
helper functions above from helpers.py instead of defining them
locally. This file contains no helper functions of its own, only
scenario orchestration.

The 8 upgrade test files with bespoke logic (the *_leader_change_from_298.py
and *_from_280_via_298.py variants, plus test_upgrade.py) are restored
to byte-identical content with main -- their only prior difference was
which module they imported the same helper functions from. The 10
thin-wrapper pair files (test_upgrade_{no_tls,ssc,ssc_single_unit,mtls,
mtls_single_unit}_from_{280,298}.py) keep their run_*_upgrade_scenario
import, now pointed at upgrade_tests_scenarios instead of the removed
upgrade_tests_helper.

Verified: tox -e integration -- --collect-only still collects 85 tests
(unchanged from before this reorganization -- confirms it's a pure
internal move with no behavioral change). git diff main -- <file>
is empty for helpers.py and all 9 bespoke-logic files. Full tox -e lint
pass clean.
