+++ tests/integration/helpers.py
@@ -173,6 +173,18 @@ def fetch_with_retry(url: str, expected_status: int = 200) -> requests.Response:
     return _fetch()
 
 
+def assert_traefik_revision(juju: jubilant.Juju, expected_revision: int) -> None:
+    """Assert traefik's deployed charm revision matches *expected_revision*.
+
+    The locally built charm reports revision ``0``; a charm refreshed to a
+    specific Charmhub revision reports that revision.
+    """
+    actual_revision = juju.status().apps[TRAEFIK_APP_NAME].charm_rev
+    assert actual_revision == expected_revision, (
+        f"Expected traefik at revision {expected_revision}, but found {actual_revision}"
+    )
+
+
 def generate_ca(tmp_path: Path) -> None:
     """Create a self-signed CA and write its certificate to disk.
 
@@ -387,11 +399,86 @@ def verify_https_on_all_units(
     return ingress_url
 
 
+def verify_http_on_all_units(
+    juju: jubilant.Juju,
+    expected_url: Optional[str] = None,
+) -> str:
+    """Assert HTTP is reachable through every traefik unit.
+
+    Returns the ingress URL that was verified so callers can assert it is
+    unchanged across an upgrade.
+    """
+    ingress_url = f"{_ingress_url(juju).rstrip('/')}/health"
+    assert ingress_url.startswith("http://"), (
+        f"expected plain HTTP proxied URL without a certificate provider, got {ingress_url!r}"
+    )
+    if expected_url is not None:
+        assert ingress_url == expected_url, (
+            f"Proxied URL changed across upgrade: {expected_url!r} -> {ingress_url!r}"
+        )
+
+    status = juju.status()
+    units = status.apps[TRAEFIK_APP_NAME].units
+
+    for unit_name, unit_status in units.items():
+        unit_ip = unit_status.address
+        logger.info("Verifying HTTP on %s (%s) -> %s", unit_name, unit_ip, ingress_url)
+        session = requests.Session()
+        session.headers["Host"] = MOCK_HOSTNAME
+        _get_with_retry(session, _url_for_unit(ingress_url, unit_ip))
+
+    return ingress_url
+
+
+def leader_unit_name(juju: jubilant.Juju, app: str = TRAEFIK_APP_NAME) -> str:
+    """Return the name of the current leader unit of *app*."""
+    for name, unit in juju.status().apps[app].units.items():
+        if unit.leader:
+            return name
+    raise AssertionError(f"no leader found for {app!r}")
+
+
 def _unit_address(juju: jubilant.Juju, unit_name: str, app: str = TRAEFIK_APP_NAME) -> str:
     """Return unit IP address for *unit_name* in *app* or raise if missing."""
     units = juju.status().apps[app].units
     assert unit_name in units, f"{unit_name} not found in {app} units"
     return units[unit_name].address
+
+
+def force_leader_change(juju: jubilant.Juju, app: str = TRAEFIK_APP_NAME) -> str:
+    """Force a leadership change by stopping the current leader's unit agent."""
+    old_leader = leader_unit_name(juju, app)
+    logger.info(
+        "Stopping the container-agent on leader %s to force a leadership change", old_leader
+    )
+    # stop-checks liveness prevents pebble from restarting the agent as unhealthy.
+    juju.ssh(old_leader, "/charm/bin/pebble", "stop-checks", "liveness", container="charm")
+    juju.ssh(old_leader, "/charm/bin/pebble", "stop", "container-agent", container="charm")
+
+    def _reelected(status: jubilant.Status) -> bool:
+        units = status.apps[app].units
+        leaders = [name for name, unit in units.items() if unit.leader]
+        return len(leaders) == 1 and leaders[0] != old_leader
+
+    try:
+        # No error= here: the old leader's agent is deliberately stopped above, so it
+        # may legitimately report "lost"/error while we wait for a new leader to be elected.
+        juju.wait(_reelected, timeout=120, delay=5)
+    except TimeoutError as exc:
+        raise AssertionError(
+            f"leadership did not move away from {old_leader} within 2 minutes"
+        ) from exc
+    new_leader = leader_unit_name(juju, app)
+    logger.info("Leadership moved from %s to %s", old_leader, new_leader)
+    # Trigger a hook on the new leader so it can react to the leadership change.
+    # Traefik currently does not observe leader-elected hook.
+    juju.config(app, {"loadbalancer_annotations": " "})
+    # Bring the old leader back: re-enable liveness checks and restart its
+    # container-agent.
+    logger.info("Restarting container-agent and liveness checks on %s", old_leader)
+    juju.ssh(old_leader, "/charm/bin/pebble", "start", "container-agent", container="charm")
+    juju.ssh(old_leader, "/charm/bin/pebble", "start-checks", "liveness", container="charm")
+    return new_leader
 
 
 def verify_https_on_unit(juju: jubilant.Juju, unit_name: str, ingress_url: str) -> None:
@@ -403,6 +490,18 @@ def verify_https_on_unit(juju: jubilant.Juju, unit_name: str, ingress_url: str) 
     session.verify = str(ca_cert_path)
     response = session.get(ingress_url, timeout=30)
     response.raise_for_status()
+
+
+def verify_http_on_unit(juju: jubilant.Juju, unit_name: str, ingress_url: str) -> None:
+    """Assert HTTP returns 200 on a specific traefik unit."""
+    assert ingress_url.startswith("http://"), (
+        f"expected plain HTTP proxied URL without a certificate provider, got {ingress_url!r}"
+    )
+    unit_ip = _unit_address(juju, unit_name)
+    logger.info("Verifying HTTP on %s (%s) -> %s", unit_name, unit_ip, ingress_url)
+    session = requests.Session()
+    session.headers["Host"] = MOCK_HOSTNAME
+    _get_with_retry(session, _url_for_unit(ingress_url, unit_ip))
 
 
 # --- Composite flows --------------------------------------------------------
@@ -439,3 +538,10 @@ def bring_up_self_signed_traefik(
     pull_ssc_ca_certificate(juju, tmp_path, ssc_app=ssc_app)
 
     return verify_https_on_all_units(juju)
+
+
+def bring_up_traefik_without_certificate_provider(juju: jubilant.Juju) -> str:
+    """Integrate the ingress requirer and verify plain HTTP on all traefik units."""
+    juju.integrate(f"{INGRESS_REQUIRER_APP_NAME}:require-ingress", TRAEFIK_APP_NAME)
+    juju.wait(all_settled, error=jubilant.any_error, delay=5, timeout=900, successes=5)
+    return verify_http_on_all_units(juju)
