+++ tests/integration/test_upgrade_mtls_from_280.py
@@ -18,7 +18,7 @@ Scenario:
 import jubilant
 import pytest
 from conftest import TRAEFIK_RESOURCES
-from upgrade_tests_helper import run_mtls_upgrade_scenario
+from upgrade_tests_scenarios import run_mtls_upgrade_scenario
 
 SOURCE_REVISION = 280
