+++ tests/integration/test_upgrade_no_tls_leader_change_from_298.py
@@ -20,8 +20,8 @@ import jubilant
 import pytest
 from conftest import TRAEFIK_APP_NAME, TRAEFIK_RESOURCES
 from constants import MOCK_HOSTNAME, NUM_TRAEFIK_UNITS, SOURCE_CHANNEL, TRAEFIK_CHARM
-from helpers import all_settled
-from upgrade_tests_helper import (
+from helpers import (
+    all_settled,
     assert_traefik_revision,
     bring_up_traefik_without_certificate_provider,
     force_leader_change,
