+++ tests/integration/test_upgrade_ssc_from_280.py
@@ -16,7 +16,7 @@ Scenario:
 import jubilant
 import pytest
 from conftest import TRAEFIK_RESOURCES
-from upgrade_tests_helper import run_ssc_upgrade_scenario
+from upgrade_tests_scenarios import run_ssc_upgrade_scenario
 
 SOURCE_REVISION = 280
