+++ tests/integration/test_upgrade_ssc_leader_change_from_298.py
@@ -24,10 +24,11 @@ from conftest import TRAEFIK_APP_NAME, TRAEFIK_RESOURCES
 from constants import MOCK_HOSTNAME, NUM_TRAEFIK_UNITS, SOURCE_CHANNEL, TRAEFIK_CHARM
 from helpers import (
     all_settled,
+    assert_traefik_revision,
     bring_up_self_signed_traefik,
+    force_leader_change,
     verify_https_on_all_units,
 )
-from upgrade_tests_helper import assert_traefik_revision, force_leader_change
 
 logger = logging.getLogger(__name__)
