+++ tests/integration/test_upgrade_ssc_single_unit_from_298.py
@@ -7,7 +7,7 @@
 import jubilant
 import pytest
 from conftest import TRAEFIK_RESOURCES
-from upgrade_tests_helper import run_ssc_single_unit_upgrade_scenario
+from upgrade_tests_scenarios import run_ssc_single_unit_upgrade_scenario
 
 SOURCE_REVISION = 298
